quantum_ml_time_series_project/src/recurrent_quantum_nn.py

import numpy as np
from qiskit import QuantumCircuit, Aer, transpile, execute
from qiskit_machine_learning.circuit.library import ZZFeatureMap
from qiskit_machine_learning.algorithms import VQC
from qiskit_machine_learning.kernels import QuantumKernel
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split

class RecurrentQuantumNN:
    def __init__(self, num_qubits, time_steps, learning_rate=0.01):
        self.num_qubits = num_qubits
        self.time_steps = time_steps
        self.learning_rate = learning_rate

        # Define the quantum circuit
        self.feature_map = ZZFeatureMap(self.num_qubits)
        self.circuit = QuantumCircuit(self.num_qubits)

        # Quantum simulator
        self.simulator = Aer.get_backend('statevector_simulator')

    def build_model(self, data):
        """Create the model and return it after training."""
        X, y = self._prepare_data(data)
        model = VQC(feature_map=self.feature_map, quantum_instance=self.simulator)
        model.fit(X, y)
        return model

    def _prepare_data(self, data):
        """Prepares time series data for recurrent quantum model."""
        X = []
        y = []
        for i in range(len(data) - self.time_steps):
            X.append(data[i:i + self.time_steps])
            y.append(data[i + self.time_steps])
        return np.array(X), np.array(y)

    def predict(self, model, data):
        """Predict the next value in the time series."""
        X, _ = self._prepare_data(data)
        predictions = model.predict(X)
        return predictions

    def evaluate(self, model, data):
        """Evaluate the model performance using mean squared error."""
        X, y_true = self._prepare_data(data)
        y_pred = model.predict(X)
        return mean_squared_error(y_true, y_pred)

    def optimize_hyperparameters(self, data, param_grid):
        """Optimizes hyperparameters using grid search or other methods."""
        best_params = {}
        best_score = float('inf')

        # Basic grid search implementation
        for params in param_grid:
            model = self.build_model(data)
            score = self.evaluate(model, data)
            if score < best_score:
                best_score = score
                best_params = params
        
        return best_params, best_score

    def save_model(self, model, path):
        """Save the quantum model to a file."""
        model.save(path)

    def load_model(self, path):
        """Load the quantum model from a file."""
        return VQC.load(path)
